# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Test project",
        "title": "Song Library",
        "contact": {
            "name": "go_rshok",
            "url": "https://t.me/go_rshok"
        },
        "version": "0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/song": {
            "get": {
                "description": "Returns song by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Get Song",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get song by name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSongResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns song",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Create song",
                "parameters": [
                    {
                        "description": "Create Song",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSongRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAndUpdateSongResponseDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete song by song name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Delete song",
                "parameters": [
                    {
                        "description": "Delete Song",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteSongRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Returns song",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Update song",
                "parameters": [
                    {
                        "description": "Update Song",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSongRequestDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Update song name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAndUpdateSongResponseDTO"
                        }
                    }
                }
            }
        },
        "/songs": {
            "get": {
                "description": "Returns songs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Get Songs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "order",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order_field",
                        "name": "order_field",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSongsResponseDTO"
                        }
                    }
                }
            }
        },
        "/text-song": {
            "get": {
                "description": "Returns song by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Get Text Song",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get text song by name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateAndUpdateSongResponseDTO": {
            "type": "object",
            "properties": {
                "song": {
                    "$ref": "#/definitions/entity.Song"
                }
            }
        },
        "dto.CreateSongRequestDTO": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "listenings": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "dto.DeleteSongRequestDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.GetSongResponseDTO": {
            "type": "object",
            "properties": {
                "song": {
                    "$ref": "#/definitions/entity.Song"
                }
            }
        },
        "dto.GetSongsResponseDTO": {
            "type": "object",
            "properties": {
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Song"
                    }
                }
            }
        },
        "dto.UpdateSongRequestDTO": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "entity.Song": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "listenings": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}